requirements.txt
flask==2.3.3
requests==2.31.0
oauthlib==3.2.2
python-dotenv==1.0.0

# psycopg2-binary is only needed for production deployment
# It's commented out for local development to avoid installation issues
# psycopg2-binary==2.9.9
